Fix serialize test failure on 64 bit architerctures.
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid@644125 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/amqp_0_10/Map.h
@@ -157,7 +157,7 @@ template <class V> typename V::result_type MapValue::apply_visitor(const V& v) {
 }
 
 template <class S> void Map::encode(S& s) const {
-    s(contentSize())(size()); // size, count
+    s(contentSize())(uint32_t(size())); // size, count
     for (const_iterator i = begin(); i != end(); ++i)
         s(i->first)(i->second); // key (type value)
 }
